# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wedding Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weddings/{weddingId}/schedule:
    get:
      operationId: listScheduleItems
      tags: [schedule]
      summary: List schedule items
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
    post:
      operationId: createScheduleItem
      tags: [schedule]
      summary: Create schedule item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleItemInput"
      responses:
        "201":
          description: Schedule item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleItem"
  /weddings/{weddingId}/schedule/{id}:
    patch:
      operationId: updateScheduleItem
      tags: [schedule]
      summary: Update schedule item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleItemInput"
      responses:
        "200":
          description: Schedule item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleItem"
    delete:
      operationId: deleteScheduleItem
      tags: [schedule]
      summary: Delete schedule item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /weddings/{weddingId}/tables:
    get:
      operationId: listTables
      tags: [seating]
      summary: List tables
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeatingTable"
    post:
      operationId: createTable
      tags: [seating]
      summary: Create table
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatingTableInput"
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatingTable"
  /weddings/{weddingId}/tables/{id}:
    patch:
      operationId: updateTable
      tags: [seating]
      summary: Update table
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatingTableInput"
      responses:
        "200":
          description: Table updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatingTable"
    delete:
      operationId: deleteTable
      tags: [seating]
      summary: Delete table
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/seat-assignments:
    get:
      operationId: listSeatAssignments
      tags: [seating]
      summary: List seat assignments
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seat assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeatAssignment"
    post:
      operationId: assignSeat
      tags: [seating]
      summary: Assign guest to table
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatAssignmentInput"
      responses:
        "201":
          description: Seat assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatAssignment"
  /weddings/{weddingId}/seat-assignments/{id}:
    delete:
      operationId: removeSeatAssignment
      tags: [seating]
      summary: Remove seat assignment
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /weddings/{weddingId}/settings:
    get:
      operationId: getIntegrationSettings
      tags: [settings]
      summary: Get integration settings
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationSettings"
    put:
      operationId: updateIntegrationSettings
      tags: [settings]
      summary: Update integration settings
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationSettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationSettings"
  /weddings/{weddingId}/settings/test-whatsapp:
    post:
      operationId: testWhatsappConnection
      tags: [settings]
      summary: Test WhatsApp connection
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /weddings/{weddingId}/settings/test-asaas:
    post:
      operationId: testAsaasConnection
      tags: [settings]
      summary: Test Asaas connection
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
...
components:
  schemas:
...
    ScheduleItem:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        responsible:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - title
        - startTime
        - sortOrder
        - createdAt
    ScheduleItemInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        responsible:
          type: ["string", "null"]
        sortOrder:
          type: integer
...
    SeatingTable:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        positionX:
          type: number
        positionY:
          type: number
        shape:
          type: string
          enum: [round, rectangular, square]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - name
        - capacity
        - positionX
        - positionY
        - shape
        - createdAt
    SeatingTableInput:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        positionX:
          type: number
        positionY:
          type: number
        shape:
          type: string
          enum: [round, rectangular, square]
    SeatAssignment:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        tableId:
          type: integer
        guestId:
          type: integer
        seatNumber:
          type: ["integer", "null"]
        guestName:
          type: ["string", "null"]
        tableName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - tableId
        - guestId
        - createdAt
    SeatAssignmentInput:
      type: object
      properties:
        tableId:
          type: integer
        guestId:
          type: integer
        seatNumber:
          type: ["integer", "null"]
      required:
        - tableId
        - guestId
...
    IntegrationSettings:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        evolutionApiUrl:
          type: ["string", "null"]
        evolutionApiKey:
          type: ["string", "null"]
        evolutionInstance:
          type: ["string", "null"]
        asaasApiKey:
          type: ["string", "null"]
        asaasEnvironment:
          type: string
          enum: [sandbox, production]
        asaasWebhookToken:
          type: ["string", "null"]
      required:
        - id
        - weddingId
        - asaasEnvironment
    IntegrationSettingsInput:
      type: object
      properties:
        evolutionApiUrl:
          type: ["string", "null"]
        evolutionApiKey:
          type: ["string", "null"]
        evolutionInstance:
          type: ["string", "null"]
        asaasApiKey:
          type: ["string", "null"]
        asaasEnvironment:
          type: string
          enum: [sandbox, production]
        asaasWebhookToken:
          type: ["string", "null"]
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...