# lib/api-zod/src/generated/types/integrationSettingsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Management API
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrationSettingsInputAsaasEnvironment } from "./integrationSettingsInputAsaasEnvironment";

export interface IntegrationSettingsInput {
  /** @nullable */
  evolutionApiUrl?: string | null;
  /** @nullable */
  evolutionApiKey?: string | null;
  /** @nullable */
  evolutionInstance?: string | null;
  /** @nullable */
  asaasApiKey?: string | null;
  asaasEnvironment?: IntegrationSettingsInputAsaasEnvironment;
  /** @nullable */
  asaasWebhookToken?: string | null;
}