# lib/api-zod/src/generated/types/seatAssignmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SeatAssignmentInput {
  tableId: number;
  guestId: number;
  /** @nullable */
  seatNumber?: number | null;
}